Refactor trace writer bundle module
